refactor: Header 컴포넌트 재분류
